--- internal/jobqueue/local_target_uploader.go
@@ -1,7 +1,0 @@
-package jobqueue
-
-import "io"
-
-type LocalTargetUploader interface {
-	UploadImage(job *Job, reader io.Reader) error
-}
